+++ processor/src/test/resources/expected/com/example/arez/ComponentWithAnnotatedParameterEventHandler_.java
@@ -14,9 +14,7 @@ import react4j.core.BaseProps;
 import react4j.core.BaseState;
 import react4j.core.ComponentConstructorFunction;
 import react4j.core.NativeAdapterComponent;
-import react4j.core.React;
 import react4j.core.ReactConfig;
-import react4j.core.ReactNode;
 
 @ArezComponent(
     type = "ComponentWithAnnotatedParameterEventHandler",
@@ -31,21 +29,6 @@ abstract class ComponentWithAnnotatedParameterEventHandler_ extends ComponentWit
 
   @Nonnull
   private final ComponentWithAnnotatedParameterEventHandler.CustomHandler _handleFoo2 = create_handleFoo2();
-
-  @Nonnull
-  static ReactNode _create() {
-    return React.createElement( TYPE );
-  }
-
-  @Nonnull
-  static ReactNode _create(@Nullable final BaseProps props) {
-    return React.createElement( TYPE, props );
-  }
-
-  @Nonnull
-  static ReactNode _create(@Nullable final BaseProps props, @Nullable final ReactNode child) {
-    return React.createElement( TYPE, props, child );
-  }
 
   @Nonnull
   private static ComponentConstructorFunction<BaseProps, BaseContext> getConstructorFunction() {