+++ processor/src/test/resources/expected/com/example/arez/ComponentWithNoAutoActionEventHandler_.java
@@ -14,9 +14,7 @@ import react4j.core.BaseProps;
 import react4j.core.BaseState;
 import react4j.core.ComponentConstructorFunction;
 import react4j.core.NativeAdapterComponent;
-import react4j.core.React;
 import react4j.core.ReactConfig;
-import react4j.core.ReactNode;
 
 @ArezComponent(
     type = "ComponentWithNoAutoActionEventHandler",
@@ -28,21 +26,6 @@ abstract class ComponentWithNoAutoActionEventHandler_ extends ComponentWithNoAut
 
   @Nonnull
   private final EventHandler.Procedure _handleFoo = create_handleFoo();
-
-  @Nonnull
-  static ReactNode _create() {
-    return React.createElement( TYPE );
-  }
-
-  @Nonnull
-  static ReactNode _create(@Nullable final BaseProps props) {
-    return React.createElement( TYPE, props );
-  }
-
-  @Nonnull
-  static ReactNode _create(@Nullable final BaseProps props, @Nullable final ReactNode child) {
-    return React.createElement( TYPE, props, child );
-  }
 
   @Nonnull
   private static ComponentConstructorFunction<BaseProps, BaseContext> getConstructorFunction() {