+++ processor/src/test/resources/expected/com/example/arez/ComponentWithProp_.java
@@ -13,9 +13,7 @@ import react4j.core.BaseProps;
 import react4j.core.BaseState;
 import react4j.core.ComponentConstructorFunction;
 import react4j.core.NativeAdapterComponent;
-import react4j.core.React;
 import react4j.core.ReactConfig;
-import react4j.core.ReactNode;
 
 @ArezComponent(
     type = "ComponentWithProp",
@@ -24,21 +22,6 @@ import react4j.core.ReactNode;
 @Generated("react4j.processor.ReactProcessor")
 abstract class ComponentWithProp_ extends ComponentWithProp {
   static final ComponentConstructorFunction<BaseProps, BaseContext> TYPE = getConstructorFunction();
-
-  @Nonnull
-  static ReactNode _create() {
-    return React.createElement( TYPE );
-  }
-
-  @Nonnull
-  static ReactNode _create(@Nullable final BaseProps props) {
-    return React.createElement( TYPE, props );
-  }
-
-  @Nonnull
-  static ReactNode _create(@Nullable final BaseProps props, @Nullable final ReactNode child) {
-    return React.createElement( TYPE, props, child );
-  }
 
   @Nonnull
   private static ComponentConstructorFunction<BaseProps, BaseContext> getConstructorFunction() {