+++ processor/src/test/resources/expected/com/example/basic/CustomPropsAndStateReactComponent_.java
@@ -7,29 +7,11 @@ import jsinterop.base.Js;
 import react4j.core.BaseContext;
 import react4j.core.ComponentConstructorFunction;
 import react4j.core.NativeAdapterComponent;
-import react4j.core.React;
 import react4j.core.ReactConfig;
-import react4j.core.ReactNode;
 
 @Generated("react4j.processor.ReactProcessor")
 class CustomPropsAndStateReactComponent_ extends CustomPropsAndStateReactComponent {
   static final ComponentConstructorFunction<CustomPropsAndStateReactComponent.Props, BaseContext> TYPE = getConstructorFunction();
-
-  @Nonnull
-  static ReactNode _create() {
-    return React.createElement( TYPE );
-  }
-
-  @Nonnull
-  static ReactNode _create(@Nullable final CustomPropsAndStateReactComponent.Props props) {
-    return React.createElement( TYPE, props );
-  }
-
-  @Nonnull
-  static ReactNode _create(@Nullable final CustomPropsAndStateReactComponent.Props props,
-      @Nullable final ReactNode child) {
-    return React.createElement( TYPE, props, child );
-  }
 
   @Nonnull
   private static ComponentConstructorFunction<CustomPropsAndStateReactComponent.Props, BaseContext> getConstructorFunction() {