+++ processor/src/test/resources/expected/com/example/event_handler/BasicHandlerComponent_.java
@@ -12,9 +12,7 @@ import react4j.core.BaseProps;
 import react4j.core.BaseState;
 import react4j.core.ComponentConstructorFunction;
 import react4j.core.NativeAdapterComponent;
-import react4j.core.React;
 import react4j.core.ReactConfig;
-import react4j.core.ReactNode;
 
 @Generated("react4j.processor.ReactProcessor")
 class BasicHandlerComponent_ extends BasicHandlerComponent {
@@ -22,21 +20,6 @@ class BasicHandlerComponent_ extends BasicHandlerComponent {
 
   @Nonnull
   private final EventHandler.Procedure _handleFoo = create_handleFoo();
-
-  @Nonnull
-  static ReactNode _create() {
-    return React.createElement( TYPE );
-  }
-
-  @Nonnull
-  static ReactNode _create(@Nullable final BaseProps props) {
-    return React.createElement( TYPE, props );
-  }
-
-  @Nonnull
-  static ReactNode _create(@Nullable final BaseProps props, @Nullable final ReactNode child) {
-    return React.createElement( TYPE, props, child );
-  }
 
   @Nonnull
   private static ComponentConstructorFunction<BaseProps, BaseContext> getConstructorFunction() {