+++ processor/src/test/resources/expected/com/example/event_handler/CustomHandlerMultipleArgsComponent_.java
@@ -11,9 +11,7 @@ import react4j.core.BaseProps;
 import react4j.core.BaseState;
 import react4j.core.ComponentConstructorFunction;
 import react4j.core.NativeAdapterComponent;
-import react4j.core.React;
 import react4j.core.ReactConfig;
-import react4j.core.ReactNode;
 
 @Generated("react4j.processor.ReactProcessor")
 class CustomHandlerMultipleArgsComponent_ extends CustomHandlerMultipleArgsComponent {
@@ -21,21 +19,6 @@ class CustomHandlerMultipleArgsComponent_ extends CustomHandlerMultipleArgsCompo
 
   @Nonnull
   private final CustomHandlerMultipleArgsComponent.CustomHandler _handleFoo = create_handleFoo();
-
-  @Nonnull
-  static ReactNode _create() {
-    return React.createElement( TYPE );
-  }
-
-  @Nonnull
-  static ReactNode _create(@Nullable final BaseProps props) {
-    return React.createElement( TYPE, props );
-  }
-
-  @Nonnull
-  static ReactNode _create(@Nullable final BaseProps props, @Nullable final ReactNode child) {
-    return React.createElement( TYPE, props, child );
-  }
 
   @Nonnull
   private static ComponentConstructorFunction<BaseProps, BaseContext> getConstructorFunction() {