+++ processor/src/test/resources/expected/com/example/lifecycle/OverrideLifecycleMethodsComponent_.java
@@ -11,29 +11,12 @@ import react4j.core.BaseProps;
 import react4j.core.BaseState;
 import react4j.core.ComponentConstructorFunction;
 import react4j.core.NativeAdapterComponent;
-import react4j.core.React;
 import react4j.core.ReactConfig;
 import react4j.core.ReactErrorInfo;
-import react4j.core.ReactNode;
 
 @Generated("react4j.processor.ReactProcessor")
 class OverrideLifecycleMethodsComponent_ extends OverrideLifecycleMethodsComponent {
   static final ComponentConstructorFunction<BaseProps, BaseContext> TYPE = getConstructorFunction();
-
-  @Nonnull
-  static ReactNode _create() {
-    return React.createElement( TYPE );
-  }
-
-  @Nonnull
-  static ReactNode _create(@Nullable final BaseProps props) {
-    return React.createElement( TYPE, props );
-  }
-
-  @Nonnull
-  static ReactNode _create(@Nullable final BaseProps props, @Nullable final ReactNode child) {
-    return React.createElement( TYPE, props, child );
-  }
 
   @Nonnull
   private static ComponentConstructorFunction<BaseProps, BaseContext> getConstructorFunction() {