+++ processor/src/test/resources/expected/com/example/nested/NestedCompleteComponent_BasicReactComponent_.java
@@ -15,9 +15,7 @@ import react4j.core.BaseProps;
 import react4j.core.BaseState;
 import react4j.core.ComponentConstructorFunction;
 import react4j.core.NativeAdapterComponent;
-import react4j.core.React;
 import react4j.core.ReactConfig;
-import react4j.core.ReactNode;
 
 @Generated("react4j.processor.ReactProcessor")
 class NestedCompleteComponent_BasicReactComponent_ extends NestedCompleteComponent.BasicReactComponent {
@@ -39,21 +37,6 @@ class NestedCompleteComponent_BasicReactComponent_ extends NestedCompleteCompone
   private static Provider<NestedCompleteComponent.BasicReactComponent> getProvider() {
     Guards.invariant( () -> null != c_provider, () -> "Attempted to create an instance of the React4j component named 'BasicReactComponent' before the dependency injection provider has been initialized. Please see the documentation at https://react4j.github.io/dependency_injection for directions how to configure dependency injection." );
     return c_provider;
-  }
-
-  @Nonnull
-  static ReactNode _create() {
-    return React.createElement( TYPE );
-  }
-
-  @Nonnull
-  static ReactNode _create(@Nullable final BaseProps props) {
-    return React.createElement( TYPE, props );
-  }
-
-  @Nonnull
-  static ReactNode _create(@Nullable final BaseProps props, @Nullable final ReactNode child) {
-    return React.createElement( TYPE, props, child );
   }
 
   @Nonnull